Liste Client pour admin

+++ controllers/deleteClientCtrl.php
@@ -1,22 +1,26 @@
 <?php
-$client = new Client();
 
-// Condition qui permet de vérifier les parametres d'URL
-if (isset($_GET['id']) && $_GET['id'] > 0){
-    
-    $client->id = htmlspecialchars($_GET['id']);
+if (isset($_SESSION['user_id']) && $_SESSION['user_id'] == $_GET['id']) {
+    $client = new Client();
 
-    $success = $client->deleteClient();
+    if (isset($_GET['id']) && $_GET['id'] > 0) {
 
-    if ($success) {
+        $client->id = htmlspecialchars($_GET['id']);
+        $success = $client->deleteClient();
 
-        // création du message de confirmation 
-        $_SESSION['message'] = 'Votre compte a bien été supprimé';
+        if ($success) {
+            $_SESSION['message'] = 'Le compte a bien été supprimé';
+            if ($client->id == $_SESSION['user_id']) {
+                session_destroy();
+            }
+            header('Location: ../index.php');
+            exit();
+        }
+    } else {
         header('Location: ../index.php');
         exit();
     }
 } else {
-    // Le else permet de renvoyer l'utilisateur si il y a autre chose qu'un type int
     header('Location: ../index.php');
     exit();
 }
